requirements.txt
aiohttp==3.6.2
alabaster==0.7.12
altgraph==0.16.1
async-timeout==3.0.1
attrs==19.3.0
Babel==2.7.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
decorator==4.4.1
discord==1.0.1
discord.py==1.2.5
docutils==0.15.2
future==0.18.2
idna==2.8
imageio==2.6.1
imageio-ffmpeg==0.3.0
imagesize==1.1.0
instagram-private-api==1.6.0
InstagramAPI==1.0.2
Jinja2==2.10.3
MarkupSafe==1.1.1
moviepy==1.0.1
multidict==4.6.1
nltk==3.4.5
numpy==1.17.4
packaging==19.2
pefile==2019.4.18
Pillow==6.2.1
preprocessing==0.1.13
proglog==0.1.9
Pygments==2.5.1
PyInstaller==3.5
pyparsing==2.4.5
pytz==2019.3
pywin32-ctypes==0.2.0
requests==2.22.0
requests-toolbelt==0.9.1
six==1.13.0
snowballstemmer==2.0.0
Sphinx==2.2.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
tqdm==4.39.0
urllib3==1.25.7
vk-api==11.6.1
websockets==7.0
yarl==1.3.0
